Partial implementation of code of P functionality for student self registration
some corrections were made to the database as I noticed some minor errors, I also needed to add another member to the attendance class. I began work on the self registration of attendance for students, as my work is dependent on that of another group member, I was unable to complete it at this time, but I have been able to complete parts. The code will check if there is a timetable event that matches the check in code, when this is confirmed, the student id , timetable event and attendance status are added to the table in database, and a message is displayed that attendance has been marked.